decouple stories and use hooks
