+++ src/js/components/DropButton/stories/Calendar.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { FormDown } from 'grommet-icons';
+
+import { Grommet, Box, Calendar, DropButton, Text } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const CalendarDropButton = () => {
+  const [date, setDate] = React.useState(undefined);
+  const [open, setOpen] = React.useState(undefined);
+
+  const onSelect = selectedDate => {
+    setDate(selectedDate);
+    setOpen(false);
+  };
+
+  return (
+    <Grommet theme={grommet}>
+      <Box align="center" pad="large">
+        <DropButton
+          open={open}
+          onClose={() => setOpen(false)}
+          onOpen={() => setOpen(true)}
+          dropContent={<Calendar date={date} onSelect={onSelect} />}
+        >
+          <Box direction="row" gap="medium" align="center" pad="small">
+            <Text>
+              {date ? new Date(date).toLocaleDateString() : 'Select date'}
+            </Text>
+            <FormDown color="brand" />
+          </Box>
+        </DropButton>
+      </Box>
+    </Grommet>
+  );
+};
+
+storiesOf('DropButton', module).add('Calendar', () => <CalendarDropButton />);
